EditPost

+++ app/components/blog-detail/DetailPostPage.js
@@ -22,16 +22,18 @@ class DetailPostPage extends React.Component {
     }
 
     handleDeletePost(){
-        const postId = this.props.post.id;
-        this.props.actions.deletePost(postId)
-            .then(() => {
-                toastr.success('Post removed');
-                //redirect
-                this.context.router.push('/app/blog');
-            })
-            .catch(error => {
-                toastr.error(error);
-            });
+        if (confirm('Do you want to delete the Post with title: ' + this.props.post.title)){
+            const postId = this.props.post.id;
+            this.props.actions.deletePost(postId)
+                .then(() => {
+                    toastr.success('Post removed');
+                    //redirect
+                    this.context.router.push('/app/blog');
+                })
+                .catch(error => {
+                    toastr.error(error);
+                });
+        }
     }
 
     render() {
@@ -42,7 +44,7 @@ class DetailPostPage extends React.Component {
                         post={this.props.post}
                     />
                     <div className="well well-sm actions">
-                        <a href="#" className="btn btn-primary"><i className="glyphicon glyphicon-edit"/> Edit</a>
+                        <Link to={`/app/post-edit/${this.props.post.id}`} className="btn btn-primary"><i className="glyphicon glyphicon-edit"/> Edit</Link>
                         <a onClick={this.handleDeletePost} className="btn btn-danger"><i className="glyphicon glyphicon-trash"/> Delete</a>
                         <Link to="/app/blog" className="btn btn-default"><i className="glyphicon glyphicon-chevron-left"/> Go Back</Link>
                     </div>
